refacto(react): moved type.ts to src
